fix(kernel): round-2 review — float32 parity + close() contract
- Top-level FLOAT columns now scan to a native float32, not a widened float64
  (N3). Thrift returns float32 for a bare FLOAT, and database/sql's asString
  formats at bit-size 32, so widening rendered CAST(0.1 AS FLOAT) as
  "0.10000000149011612" vs Thrift's "0.1". The round-1 M2 fix covered only the
  nested path; this closes the top-level scalar arm. Parity test gains a
  top-level FLOAT case at a non-exactly-representable value, and the e2e float
  case moves from 1.5 (exactly representable, masked the bug) to 0.1.
- close() reports closed=false for a handle-less op (N9). The bound-params error
  path returns &kernelOp{} with no handle; conn.ExecContext closes it
  unconditionally, and the previous unconditional true recorded a phantom
  CLOSE_STATEMENT for a statement that never reached the server. Now matches the
  backend.Operation contract (closed=false when there was no handle).

Tests: add float32_native scalar case, top-level FLOAT to the parity query, and
TestExecuteRejectsParams (asserts non-nil op, closed=false, AffectedRows 0).

+++ internal/backend/kernel/rows.go
@@ -215,7 +215,11 @@ func scanCell(col arrow.Array, row int, loc *time.Location) (driver.Value, error
 	case *array.Uint64:
 		return int64(c.Value(row)), nil
 	case *array.Float32:
-		return float64(c.Value(row)), nil
+		// Return the native float32, NOT a widened float64: the Thrift path returns
+		// a float32 driver.Value for a bare FLOAT column, and database/sql's
+		// asString formats it at bit-size 32 — so widening here would render
+		// CAST(0.1 AS FLOAT) as "0.10000000149011612" vs Thrift's "0.1".
+		return c.Value(row), nil
 	case *array.Float64:
 		return c.Value(row), nil
 	case *array.String:
